Use fallback messaging mechanism on non-systemd systems
When systemd is not available we cannot use journald as the default
event communication channel. In this case it falls back to a file to
communicate the state of the server to the client so the client knows
that the server is done.
The state is persistent even if the server has exited.

Co-authored-by: Didier Roche <didrocks@ubuntu.com>